+++ R/query.R
@@ -695,9 +695,9 @@ docdb_query.src_sqlite <- function(src, key, query, ...) {
       length(fldQ$includeFields) == 1L &&
       fldQ$includeFields == "_id") {
 
-    statement <- insObj('
-    SELECT DISTINCT _id
-    FROM "/** key **/" GROUP BY _id;')
+    statement <- paste0(
+      "SELECT DISTINCT _id ",
+      "FROM \"", key, "\" GROUP BY _id;")
 
     return(DBI::dbGetQuery(
       conn = src$con,
@@ -1001,7 +1001,8 @@ docdb_query.src_postgres <- function(src, key, query, ...) {
   # - existsFields
   fldQ$existsFields <- paste0(
     sprintf("jsonb_path_exists(json, '$.\"%s\"')",
-            gsub('[.]', '"."', fldQ$includeFields[fldQ$includeFields != "_id"])), collapse = " OR ")
+            gsub('[.]', '"."', fldQ$includeFields[fldQ$includeFields != "_id"])),
+    collapse = " OR ")
   if (fldQ$existsFields == "") fldQ$existsFields <- "TRUE"
 
 
@@ -1016,6 +1017,8 @@ docdb_query.src_postgres <- function(src, key, query, ...) {
   # - adapt sql to postgresql functions
   for (i in fldQ$queryPaths[fldQ$queryPaths != "_id"]) {
 
+    # Checks whether the JSON path returns any item for the specified JSON value
+    # jsonb_path_exists('{"a":[1,2,3,4,5]}', '$.a[*] ? (@ >= $min && @ <= $max)', '{"min":2, "max":4}')
     fldQ$queryCondition <- gsub(
       paste0("(?<![.])(\"", i, "\" [INOTREGXP=!<>']+ .+?)( AND | NOT | OR |\\)*$)"),
       "jsonb_path_exists(json, \'$[*] ? (@.\\1)')\\2", # keep $[*]
@@ -1439,6 +1442,240 @@ docdb_query.src_duckdb <- function(src, key, query, ...) {
 }
 
 
+#' @export
+docdb_query.src_mariadb <- function(src, key, query, ...) {
+
+  # handle parameters
+  if (query == "") query <- "{}"
+  query <- jsonlite::minify(query)
+  params <- list(...)
+  n <- -1L
+  if (!is.null(params$limit)) n <- params$limit
+
+  # digest and mangle
+  fldQ <- digestFields(f = params$fields, q = query)
+
+  # - - - - - - - - -
+
+  # early return if only _id is requested
+  if (query == "{}" &&
+      length(fldQ$includeFields) == 1L &&
+      fldQ$includeFields == "_id") {
+
+    statement <- paste0(
+      "SELECT DISTINCT _id ",
+      "FROM `", key, "` GROUP BY _id;")
+
+    return(DBI::dbGetQuery(
+      conn = src$con,
+      n = n, statement = statement))
+
+  }
+
+  # - - - - - - - - -
+
+  # fields
+
+  # TODO
+  # MariaDB JSON_VALUE implicitly converts the number to a string
+  # SQL will often implicitly convert it back to a number
+
+  # - includeFields
+
+  # handling that mariadb JSON_EXTRACT always returns array, but
+  # is needed since JSON_VALUE does not return 1-element arrays
+
+  fldQ$selectFields <- gsub('[.]', '"."', c(
+    fldQ$includeFields, fldQ$queryFields))
+
+  for (i in seq_along(fldQ$selectFields)) {
+
+    # do not mangle _id
+    if (fldQ$selectFields[i] == "_id") next
+
+    # compose similar to below for queryCondition
+    fldQ$selectFields[i] <- paste0(
+      '(CASE JSON_TYPE(JSON_EXTRACT(`/** key **/`.json, \'$."',
+      gsub('"."', '"**."', sub('^(.+)"[.]".+?$', "\\1", fldQ$selectFields[i])), '"\')) ',
+      "WHEN 'ARRAY' ",
+      'THEN JSON_EXTRACT(`/** key **/`.json, \'$."', gsub('"[.]"', '"**."', fldQ$selectFields[i]), '"\') ',
+      'ELSE JSON_VALUE(`/** key **/`.json, \'$."', fldQ$selectFields[i], '"\') ',
+      "END) ")
+
+  }
+
+  # concat
+  fldQ$selectFields <- paste0(
+    unique(sprintf(
+      "'%s', %s",
+      c(fldQ$includeFields, fldQ$queryFields),
+      fldQ$selectFields)),
+    collapse = ", ")
+
+  # handle _id
+  if (length(fldQ$includeFields)) {
+    if (!any("_id" == fldQ$excludeFields) &&
+        !any("_id" == fldQ$includeFields)) fldQ$selectFields <- paste0(
+          '\'_id\', _id, ', fldQ$selectFields)
+    fldQ$selectFields <- paste0('JSON_OBJECT(', fldQ$selectFields, ')')
+  } else {
+    fldQ$selectFields <-
+      "CONCAT('{\"_id\": \"', _id, '\", ', TRIM(LEADING '{' FROM json))"
+  }
+
+  # - existsFields
+
+  fldQ$existsFields <- paste0(
+    sprintf("JSON_EXISTS(json, '$.\"%s\"')",
+            gsub('[.]', '"[0]."', c(
+              fldQ$includeFields[fldQ$includeFields != "_id"],
+              fldQ$queryFields[fldQ$queryFields != "_id"]
+            ))),
+    collapse = " OR ")
+  if (fldQ$existsFields == "") fldQ$existsFields <- "TRUE"
+
+  # - - - - - - - - -
+
+  # query
+
+  # - no query
+  if (!length(fldQ$queryCondition)) fldQ$queryCondition <- "TRUE"
+
+  # adapt all
+  fldQ$queryCondition <- gsub(" == ", " = ", fldQ$queryCondition)
+  fldQ$queryCondition <- gsub('"_id"', "_id", fldQ$queryCondition)
+
+  # - "tags":["aliqua","consectetur","commodo","velit","cupidatat","duis","dolore"]
+  #          ^ start of the string compared against regular expression
+  #   thus cover this and scalar strings by double quotes as alternative boundaries
+  fldQ$queryCondition <- stringi::stri_replace_all_fixed(fldQ$queryCondition, "'^", "'(\"|^)")
+  fldQ$queryCondition <- stringi::stri_replace_all_fixed(fldQ$queryCondition, "$'", "(\"|$)'")
+
+  # - transform query to use relevant json function
+  for (i in fldQ$queryPaths[fldQ$queryPaths != "_id"]) {
+
+      # complicated handling as mariadb JSON_EXTRACT always returns array
+      # but is needed since JSON_VALUE does not return 1-element arrays
+      fldQ$queryCondition <- gsub(
+        paste0("(\"", i, "\") ([INOTREGXP=!<>']+ .+?)( AND | NOT | OR |\\)*$)"),
+        # compose conditional statement
+        paste0(
+          '(CASE JSON_TYPE(JSON_EXTRACT(`/** key **/`.json, \'$."',
+          gsub(
+            '"."', '"**."',
+            # remove last path element
+            sub('^(.+)"[.]".+?$', "\\1", i)
+          ), '"\')) ',
+          "WHEN 'ARRAY' ",
+          'THEN JSON_EXTRACT(`/** key **/`.json, \'$."', gsub('"[.]"', '"**."', i), '"\') ',
+          'ELSE JSON_VALUE(`/** key **/`.json, \'$."', i, '"\') ',
+          "END) ",
+          "\\2  \\3"),
+        fldQ$queryCondition
+      )
+
+      # special handling of IN
+      ins <- stringi::stri_extract_all_regex(
+        fldQ$queryCondition, paste0("'\\$.\"", i, "\"'\\) END\\) IN \\((.+?)\\)"))[[1]]
+
+      if (!is.na(ins)) {
+
+        isString <- grepl("\\('", ins) # "5, 4" or "'a ,b', 'd, f'"
+        ii <- stringi::stri_replace_all_regex(ins, "(.+?) IN \\((.+?)\\)", "$2")
+        ix <- stringi::stri_replace_all_regex(ins, "(.+?) IN \\((.+?)\\)", "$1")
+        ii <- trimws(ii)
+        ii <- stringi::stri_split_regex(ii, ifelse(isString, "', ", ", "))[[1]]
+        ii <- stringi::stri_replace_all_regex(ii, "^'|'$", "") # start or end delimiters
+        # strings to be compared at their full length, because IN is about full elements
+        if (any(is.na(suppressWarnings(as.numeric(ii))))) ii <- paste0('(\\["|,"|^)', ii, '(",|"\\]|$)')
+
+        ii <- paste0(ix, " REGEXP '", paste0("(", ii, ")", collapse = "|"), "'")
+
+        fldQ$queryCondition <- stringi::stri_replace_all_fixed(
+          fldQ$queryCondition, ins, ii, vectorize_all = TRUE
+        )
+      }
+
+  }
+
+  # - - - - - - - - -
+
+  # early return if listfields
+  if (!is.null(params$listfields)) {
+
+    # parameter use
+    limit <- ""
+    if (n != -1L) limit <- paste0("LIMIT ", n)
+
+    # statement
+    statement <- insObj("
+    WITH RECURSIVE json_keys AS (
+    ( SELECT
+         _id,
+         CAST('$' AS VARCHAR(16383)) AS key_path,
+         j.key_name AS extracted_key,
+         JSON_EXTRACT(json, CONCAT('$.', j.key_name)) AS sub_json
+     FROM
+         `/** key **/`,
+         JSON_TABLE(
+             JSON_KEYS(json),
+             '$[*]' COLUMNS (key_name VARCHAR(16383) PATH '$')
+         ) AS j
+     WHERE /** fldQ$queryCondition **/
+     /** limit **/
+    )
+    UNION ALL
+    SELECT
+        e._id,
+        CAST(CONCAT(e.key_path, '.', e.extracted_key) AS VARCHAR(16383)) AS key_path,
+        j.key_name AS extracted_key,
+        JSON_EXTRACT(e.sub_json, CONCAT('$.', j.key_name)) AS sub_json
+    FROM
+        json_keys e,
+        JSON_TABLE(
+            JSON_KEYS(e.sub_json),
+            '$[*]' COLUMNS (key_name VARCHAR(16383) PATH '$')
+        ) AS j
+    WHERE JSON_TYPE(e.sub_json) = 'OBJECT')
+    SELECT DISTINCT TRIM(LEADING '$.' FROM CONCAT(key_path, '.', extracted_key)) AS `key`
+    FROM json_keys
+    ORDER BY `key`;")
+
+    # debug
+    if (options()[["verbose"]]) {
+      message("\nSQL: ", statement)
+    }
+
+    # process
+    return(sort(DBI::dbGetQuery(
+      conn = src$con,
+      statement = statement,
+      n = -1L)[["key"]]))
+  }
+
+  # - - - - - - - - -
+
+  # statement
+  statement <- insObj('
+    SELECT
+    /** fldQ$selectFields **/
+    AS json FROM `/** key **/`
+    WHERE  (
+    /** fldQ$existsFields **/
+    ) AND (
+    /** fldQ$queryCondition **/
+    );')
+
+  # regular processing
+  return(processDbGetQuery(
+    getData = 'paste0(DBI::dbGetQuery(conn = src$con,
+               statement = statement, n = n)[["json"]], "")',
+    excludeFields = fldQ$excludeFields
+  ))
+
+}
+
+
 ## helpers --------------------------------------
 
 
@@ -1459,8 +1696,8 @@ processDbGetQuery <- function(
 
   # debug
   if (options()[["verbose"]]) {
-    eval.parent(parse(text = 'if (exists("statement")) message("\nSQL: ", statement, "\n")'))
-    eval.parent(parse(text = 'if (exists("query")) message("\nDB: ", query, "\n")'))
+    eval.parent(parse(text = 'if (exists("statement")) message("\nSQL: ", statement)'))
+    eval.parent(parse(text = 'if (exists("query")) message("\nQUERY: ", query, "\n")'))
   }